refactor(mobile): Rename user role to passenger
