+++ mobile/app/src/main/java/com/ognjen/fleetforge/fragments/passenger/PassengerProfile.java
@@ -1,4 +1,4 @@
-package com.ognjen.fleetforge.fragments.user;
+package com.ognjen.fleetforge.fragments.passenger;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -17,10 +17,10 @@ import com.ognjen.fleetforge.activities.MainActivity;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link UserProfile#newInstance} factory method to
+ * Use the {@link PassengerProfile#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class UserProfile extends Fragment {
+public class PassengerProfile extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -32,7 +32,7 @@ public class UserProfile extends Fragment {
     private String mParam2;
     private AuthManager authManager;
 
-    public UserProfile() {
+    public PassengerProfile() {
         // Required empty public constructor
     }
 
@@ -42,11 +42,11 @@ public class UserProfile extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment UserProfile.
+     * @return A new instance of fragment PassengerProfile.
      */
     // TODO: Rename and change types and number of parameters
-    public static UserProfile newInstance(String param1, String param2) {
-        UserProfile fragment = new UserProfile();
+    public static PassengerProfile newInstance(String param1, String param2) {
+        PassengerProfile fragment = new PassengerProfile();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -66,7 +66,7 @@ public class UserProfile extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_user_profile, container, false);
+        View view = inflater.inflate(R.layout.fragment_passenger_profile, container, false);
         Button logout=view.findViewById(R.id.btn_logout);
         logout.setOnClickListener(v -> {authManager.logout();
